refactored contig -> moved to main crate
