+++ .gitignore
@@ -1,4 +1,5 @@
 data
+!src/data
 benchmark.sh
 out.tsv
 geobacillus-plasmids*
