+++ src/data/methylation.rs
@@ -1,0 +1,22 @@
+#[derive(Debug, Clone, PartialEq, Eq, Copy)]
+pub struct MethylationCoverage {
+    n_modified: u32,
+    n_valid_cov: u32,
+}
+
+impl MethylationCoverage {
+    pub fn new(n_modified: u32, n_valid_cov: u32) -> Result<Self> {
+        if n_modified > n_valid_cov {
+            bail!(
+                "Invalid coverage: n_valid_cov ({}) cannot be less than n_modified ({})",
+                n_valid_cov,
+                n_modified
+            )
+        }
+
+        Ok(Self {
+            n_modified,
+            n_valid_cov,
+        })
+    }
+}
